refactor: move scripts de queimadas para servico
